stages/stage0_hello/pyproject.toml
[project]
name = "stage0-hello"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.2",
    "google-genai>=1.29.0",
]

[tool.uv]
package=true

[project.scripts]
server="server.server:main"
client="client.client:main"
